Refactor: move Command::GetOrCreateReflog() into base class
